Deepfake-Final-main/requirements.txt
streamlit
tensorflow
Pillow
numpy
librosa
fpdf
gdown